Retain insertion order for test case started events
